Implement streaming mode for stdin input
Previously, even when using stdin as input (-f -), the entire wav file is read into memory
and the log mel spectrogram is calculated for the entire input before the first inference.

This commit adds a streaming mode, in which a separate reader thread is spawned to read from
stdin. The reader thread continuously reads audio data in 1-second chunks and passes it to the
main thread. Once the main thread has received enough audio data (10-30 seconds), it computes
the log mel spectrogram only for that audio segment, and then runs inference on it.

After decoding, the processing window is advanced according to the decoder's final timestamp.
To avoid finishing in the middle of a word, the decoder usually stops slightly before the end
of the audio segment being processed. Therefore a small amount of audio data usually needs to be
copied from the end of the processed audio segment to the beginning of the next audio segment.

Streaming mode makes it possible to process a stream (from a webcast, or an app like OBS, etc.)
and see the text output as each audio segment is processed.
However, there are some limitations:
- Input audio must be raw samples (i.e. without WAV headers). WAV format does not seem to be well suited
  for streaming, and none of the libraries I could find had support for reading from a stream or in chunks.
- Input format is the same as used in Whisper's internal processing: 32bit float, 16kHz, 1 channel.
  Converting from other formats is possible with ffmpeg: -f f32le -vn -acodec pcm_f32le -ar 16000 -ac 1
- There is a processing delay of at least 10 seconds (plus the time it takes to run inference on each
  audio segment). If inference speed is only slightly faster than realtime, the delay is close to 30 seconds.
- If output to file is enabled, the file is currently written all at once at the end of processing.
- Several other command line options are not yet fully compatible with streaming mode.